tasks_api/urls.py
from django.urls import path, include
from rest_framework import routers
from tasks_api import views

router = routers.DefaultRouter()
router.register(r'users', views.UserViewSet)
router.register(r'groups', views.GroupViewSet)
router.register(r'members', views.MemberViewSet)
router.register(r'teams', views.TeamViewSet, basename='team')

# Wire up our API using automatic URL routing.
# Additionally, we include login URLs for the browsable API.



urlpatterns = [
    path('', include(router.urls)),
    path('api-auth/', include('rest_framework.urls', namespace='rest_framework'))

]