requirements.txt
psycopg2-binary
psycopg2
djangorestframework
markdown       # Markdown support for the browsable API.
django-filter  # Filtering support
django-cors-headers